// packages/easy-fm/src/types.ts
export interface Script {
    name: string;
    parameter: string;
}
// ...
export interface ScriptResult {
    scriptError?: FMError;
    scriptResult?: string;
}